Se retocaron un par de cosas en la funcion para planificar procesos, y ya planifica por HRRN.

+++ funciones.h
@@ -87,7 +87,7 @@
 	void cargarProcesoCola(t_list* listaUtilizar, t_queue* colaUtilizar, int socketProcesoConsultar);
 	void cargarProcesoLista(t_list* listaUtilizar, t_list* listaProcesar, int socketProcesoCargar);
 
-	int estimarRafaga(int estimacionAnterior, int rafagaAnterior, int alfa);
+	float estimarRafaga(float estimacionAnterior, int rafagaAnterior, int alfa);
 	int countParametrosConsola(char * string);
 	void listarRecursosBloqueados(t_list* listaClavesBloqueadasRequeridas, char* key);
 	void liberarRecursosProceso(t_dictionary * dictionario, char* nombreProceso);
